+++ NEWS.md
@@ -1,3 +1,15 @@
+# comat 0.6.0
+
+* Adds normalization
+
+# comat 0.5.0
+
+* Rewrites `get_incove()` internals into rcpp (`rcpp_get_incove()`)
+
+# comat 0.4.0
+
+* Merged `rcpp_get_wecove()` and `rcpp_get_cove()` into `rcpp_get_vec()`
+
 # comat 0.3.0
 
 * Added a `NEWS.md` file to track changes to the package.